app/requirements-bak.txt
amqp==2.6.1
asent==0.6.0
asgiref==3.5.2
Babel==2.11.0
billiard==3.6.4.0
blackstone==0.1.14
blis==0.7.9
boto3==1.26.16
botocore==1.29.16
catalogue==2.0.8
celery==4.4.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
confection==0.0.3
conllu==4.5.2
cryptography==38.0.3
cycler==0.11.0
cymem==2.0.7
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==3.2.16
django-allauth==0.51.0
django-cors-headers==3.13.0
django-sse-wrapper==0.0.4
django-storages==1.13.1
djangorestframework==3.14.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.1/en_core_web_sm-3.4.1-py3-none-any.whl
fonttools==4.38.0
gunicorn==20.1.0
idna==3.4
image==1.5.33
importlib-metadata==4.13.0
Jinja2==3.1.2
jmespath==1.0.1
kiwisolver==1.4.4
kombu==4.6.11
langcodes==3.3.0
MarkupSafe==2.1.1
matplotlib==3.5.3
murmurhash==1.0.9
natsort==8.2.0
numpy==1.21.6
oauthlib==3.2.2
packaging==21.3
pandas==1.3.5
pathy==0.10.0
pdf2image==1.16.0
Pillow==9.3.0
preshed==3.0.8
psycopg2-binary==2.9.5
pycparser==2.21
pydantic==1.10.2
PyJWT==2.6.0
pypandoc-binary==1.10
pyparsing==3.0.9
PyPDF2==2.11.2
pytesseract==0.3.10
python-dateutil==2.8.2
python-poppler==0.3.0
python3-openid==3.2.0
pytz==2022.6
redis==3.4.1
requests==2.28.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
six==1.16.0
smart-open==5.2.1
spacy==3.4.3
spacy-legacy==3.0.10
spacy-loggers==1.0.3
spacyfishing==0.1.8
spacyopentapioca==0.1.6
sqlparse==0.4.3
srsly==2.4.5
sse==1.2
thinc==8.1.5
tornado==6.2
tqdm==4.64.1
typer==0.7.0
typing_extensions==4.1.1
urllib3==1.26.13
vine==1.3.0
wasabi==0.10.1
zipp==3.11.0
